pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "wave-insitu"
version = "0.1.0"
description = "Tools for processing and visualizing in-situ ocean wave observations"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Edouard Gauvrit", email = "edouard.gauvrit@ifremer.fr"},
    {name = "UMR-LOPS"}
]
requires-python = ">=3.9"
keywords = ["oceanography", "waves", "visualization", "insitu", "saildrone"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: Oceanography",
]
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "xarray>=0.18.0",
    "netCDF4>=1.5.0",
    "folium>=0.12.0",
    "cartopy>=0.20.0",
    "matplotlib>=3.3.0",
    "branca>=0.4.0",
    "pyyaml>=5.4.0",
]

[project.urls]
Homepage = "https://github.com/umr-lops/wave-insitu"
Documentation = "https://github.com/umr-lops/wave-insitu#readme"
Repository = "https://github.com/umr-lops/wave-insitu"
Issues = "https://github.com/umr-lops/wave-insitu/issues"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
    "isort>=5.0",
]

[tool.setuptools]
packages = ["wave_insitu", "wave_insitu.loaders", "wave_insitu.visualization"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
